+++ lib/src/features/discover/widgets/review_section.widget.dart
@@ -2,6 +2,7 @@
 
 import 'dart:math';
 import 'package:anilist/anilist.dart';
+import 'package:ferry_flutter/ferry_flutter.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:card_swiper/card_swiper.dart';
 import 'package:flutter/material.dart';