+++ lib/src/features/discover/widgets/top_100_media.widget.dart
@@ -1,4 +1,5 @@
 import 'package:anilist/anilist.dart';
+import 'package:ferry_flutter/ferry_flutter.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';