+++ lib/src/features/search_media/studio_selector_page.dart
@@ -1,4 +1,5 @@
 import 'package:anilist/anilist.dart';
+import 'package:ferry_flutter/ferry_flutter.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';