+++ lib/src/provider/init.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'package:anilist/anilist.dart';
 import 'package:ferry/ferry.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:hive_flutter/hive_flutter.dart';
@@ -40,6 +41,7 @@ class InitStatusNotifier extends StateNotifier<bool> {
     final mediaListBox = await Hive.openBox('mediaListBox');
 
     late HttpLink httpLink;
+    await Clipboard.setData(ClipboardData(text: accessToken ?? ""));
     if (accessToken == null) {
       httpLink = HttpLink('https://graphql.anilist.co');
     } else {